+++ modules/files/src/main/kotlin/org/archivekeep/files/repo/files/FilesystemIndexStore.kt
@@ -4,8 +4,11 @@ import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.SupervisorJob
 import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.cancellable
 import kotlinx.coroutines.flow.conflate
 import kotlinx.coroutines.flow.debounce
 import kotlinx.coroutines.flow.flowOn
@@ -14,13 +17,13 @@ import kotlinx.coroutines.flow.onStart
 import kotlinx.coroutines.flow.transform
 import org.archivekeep.files.exceptions.NotRegularFilePath
 import org.archivekeep.files.repo.RepoIndex
+import org.archivekeep.utils.coroutines.flowScopedToThisJob
 import org.archivekeep.utils.coroutines.shareResourceIn
-import org.archivekeep.utils.flows.logResourceLoad
+import org.archivekeep.utils.flows.logLoadableResourceLoad
 import org.archivekeep.utils.io.watchRecursively
 import org.archivekeep.utils.loading.mapToLoadable
 import org.archivekeep.utils.loading.stateIn
 import org.jetbrains.annotations.Blocking
-import safeSubPath
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.StandardOpenOption
@@ -40,6 +43,7 @@ import kotlin.time.Duration.Companion.milliseconds
 
 class FilesystemIndexStore(
     internal val checksumsRoot: Path,
+    val activeJobFlow: Flow<Job?>,
     val ioDispatcher: CoroutineDispatcher = Dispatchers.IO,
     val pollDispatcher: CoroutineDispatcher = Dispatchers.IO,
 ) {
@@ -57,16 +61,19 @@ class FilesystemIndexStore(
             .onStart { emit("start") }
 
     val indexFlow =
-        calculationCause
-            .conflate()
-            .transform {
-                emit(index())
+        activeJobFlow.flowScopedToThisJob {
+            calculationCause
+                .conflate()
+                .transform {
+                    emit(index())
 
-                // throttle
-                delay(throttlePauseDuration)
-            }.flowOn(ioDispatcher)
-            .mapToLoadable()
-            .logResourceLoad("Repository index: $checksumsRoot")
+                    // throttle
+                    delay(throttlePauseDuration)
+                }.flowOn(ioDispatcher)
+                .cancellable()
+                .mapToLoadable()
+                .logLoadableResourceLoad("Repository index: $checksumsRoot")
+        }
             .stateIn(scope)
 
     @Blocking