+++ modules/files/src/main/kotlin/org/archivekeep/files/repo/files/util.kt
@@ -1,3 +1,5 @@
+package org.archivekeep.files.repo.files
+
 import org.archivekeep.files.exceptions.MaliciousPath
 import org.archivekeep.files.exceptions.NotNormalizedPath
 import java.nio.file.Path